refactor(i18n): restructure the package files' structure for better modularity with shared namespaces
