+++ packages/i18n/src/apps/web/types.d.ts
@@ -1,6 +1,6 @@
 // This file is automatically generated by i18next-cli, because it was not existing. You can edit it based on your needs: https://www.i18next.com/overview/typescript#custom-type-options
 
-import UiResources from "../../ui/resources.ts";
+import UiResources from "../../shared/ui/resources.ts";
 import Resources from "./resources.ts";
 
 declare module "i18next" {
