+++ packages/i18n/src/shared/locales/fr/ui.json
