+++ packages/i18n/src/shared/ui/i18next.config.ts
@@ -4,13 +4,13 @@ export default defineConfig({
   locales: ["fr", "en"],
   extract: {
     input: ["../ui/src/**/*.{ts,tsx}"],
-    output: "./src/ui/locales/{{language}}/{{namespace}}.json",
+    output: "./src/shared/locales/{{language}}/{{namespace}}.json",
     defaultNS: "ui",
   },
   types: {
-    input: "./src/ui/locales/fr/*.json",
-    output: "./src/ui/types.d.ts",
+    input: "./src/shared/locales/fr/ui.json",
+    output: "./src/shared/ui/types.d.ts",
     enableSelector: false,
-    resourcesFile: "./src/ui/resources.ts",
+    resourcesFile: "./src/shared/ui/resources.ts",
   },
 });
