+++ packages/i18n/src/shared/ui/resources.ts
